make sure the audio stream is allocated (and not finishing) before queueing data, otherwise, just drop the audio buffer audio for 0.5 seconds
git-svn-id: http://svn.mangler.org/mangler/trunk@84 f654edcb-a5fb-4f28-a0e7-68c327795dae

+++ src/mangler.h
@@ -61,7 +61,7 @@ class Mangler
         Glib::RefPtr<Gtk::StatusIcon>       statusIcon;
         ManglerChannelTree                  *channelTree;
         ManglerNetwork                      *network;
-        std::map<uint16_t, ManglerAudio* > audio;
+        std::map<uint32_t, ManglerAudio* >  audio;
         ManglerSettings                     *settings;
 
         Glib::Thread                        *networkThread;
